+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to Newspaparr will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.4] - 2025-08-07
+
+### Fixed
+- Remove duplicate renewal start logging
+- Ensure renewal block appears first in logs
+- Reduce verbosity of initialization logs
+
 ## [0.5.3] - 2025-08-07
 
 ### Fixed